+++ av1/common/convolve.h
@@ -121,8 +121,8 @@ static INLINE ConvolveParams get_conv_params_wiener(int bd) {
   return conv_params;
 }
 
-void av1_highbd_convolve_2d_facade(const uint8_t *src8, int src_stride,
-                                   uint8_t *dst, int dst_stride, int w, int h,
+void av1_highbd_convolve_2d_facade(const uint16_t *src8, int src_stride,
+                                   uint16_t *dst, int dst_stride, int w, int h,
                                    const InterpFilterParams *interp_filters[2],
                                    const int subpel_x_qn, int x_step_q4,
                                    const int subpel_y_qn, int y_step_q4,
