+++ av1/common/x86/highbd_txfm_utility_sse4.h
@@ -126,7 +126,7 @@ typedef void (*fwd_transform_1d_sse4_1)(__m128i *in, __m128i *out, int bit,
                                         const int num_cols);
 
 void av1_highbd_inv_txfm2d_add_universe_sse4_1(const int32_t *input,
-                                               uint8_t *output, int stride,
+                                               uint16_t *output, int stride,
                                                TX_TYPE tx_type, TX_SIZE tx_size,
                                                int eob, const int bd);
