+++ av1/encoder/hash_motion.h
@@ -91,7 +91,7 @@ int av1_hash_is_vertical_perfect(const YV12_BUFFER_CONFIG *picture,
                                  int block_size, int x_start, int y_start);
 
 void av1_get_block_hash_value(IntraBCHashInfo *intrabc_hash_info,
-                              const uint8_t *y_src, int stride, int block_size,
+                              const uint16_t *y_src, int stride, int block_size,
                               uint32_t *hash_value1, uint32_t *hash_value2);
 
 #ifdef __cplusplus
